[feenkcom/gtoolkit#3934] Add GtVmMemoryManagement.
Extend the base functionality for managing object space in the VM.

This adds minOldSpaceForGc, which specifies the minimum size old space should grow to before garbage collecting. In combination with the fullGcRatio, this can allow code that would otherwise trigger a large number of full GCs to complete with a single full GC at the end.

This commit adds the functionality, but is not yet enabled.